requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.7.0
async-timeout==4.0.3
attrs==24.3.0
backoff==2.2.1
beautifulsoup4==4.12.3
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.0
dataclasses-json==0.5.14
emoji==2.14.0
eval_type_backport==0.2.2
exceptiongroup==1.2.2
faiss-cpu==1.7.4
filetype==1.2.0
frozenlist==1.5.0
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.1.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
importlib-metadata==6.11.0
Jinja2==3.1.5
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.3.14
langchain-community==0.3.14
langchain-core==0.3.29
langchain-text-splitters==0.3.4
langdetect==1.0.9
langsmith==0.2.10
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.3
mdurl==0.1.2
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.20.1
nest-asyncio==1.6.0
nltk==3.9.1
numpy==1.26.4
olefile==0.47
openai==0.27.0
orjson==3.10.13
packaging==23.2
pandas==2.2.3
pickle-mixin==1.0.2
Pillow==9.5.0
propcache==0.2.1
protobuf==4.25.5
psutil==6.1.1
pyarrow==18.1.0
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.7.1
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
Pympler==1.1
pypdf==5.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-iso639==2024.10.22
python-magic==0.4.27
python-oxmsg==0.0.1
pytz==2024.2
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.2
RapidFuzz==3.11.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==1.4.54
streamlit==1.24.0
tenacity==8.5.0
tiktoken==0.4.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==4.3.1
unstructured==0.16.11
unstructured-client==0.28.1
urllib3==2.3.0
validators==0.34.0
webencodings==0.5.1
wrapt==1.17.0
yarl==1.18.3
zipp==3.21.0
